requirements.txt
streamlit
langchain
langchain-core
langchain-groq
langgraph
langchain-community
python-dotenv
pydantic
requests
pyyaml
pandas
plotly
